# src/Tools/Refitter.Generator/OpenApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Trend.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/News/GetLatestNews": {
      "get": {
        "tags": [
          "News"
        ],
        "operationId": "GetLatestNews",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/News/Deactivate/{articleId}": {
      "get": {
        "tags": [
          "News"
        ],
        "operationId": "Deactivate",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/News/Activate/{articleId}": {
      "get": {
        "tags": [
          "News"
        ],
        "operationId": "Activate",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/News/GetLatestCryptoNews": {
      "get": {
        "tags": [
          "News"
        ],
        "operationId": "GetLatestCryptoNews",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/News/GetLatestStockNews": {
      "get": {
        "tags": [
          "News"
        ],
        "operationId": "GetLatestStockNews",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/News/GetLatestForexNews": {
      "get": {
        "tags": [
          "News"
        ],
        "operationId": "GetLatestForexNews",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "title": "List<ArticleResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/SearchWord/Deactivate/{id}": {
      "delete": {
        "tags": [
          "SearchWord"
        ],
        "operationId": "Deactivate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/SearchWord/Activate/{id}": {
      "get": {
        "tags": [
          "SearchWord"
        ],
        "operationId": "Activate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArticleResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "pageSource": {
            "type": "string",
            "nullable": true
          },
          "typeName": {
            "type": "string",
            "nullable": true
          },
          "typeId": {
            "type": "integer",
            "format": "int32"
          },
          "searchWordId": {
            "type": "string",
            "nullable": true
          },
          "searchWord": {
            "type": "string",
            "nullable": true
          },
          "searchWordImage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ErrorResponseModel": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://localhost:8080/realms/PortfolioRealm/protocol/openid-connect/auth",
            "scopes": { }
          }
        }
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}