# src/Web/admin-dashboard/apps/trend-dashboard/src/app/shared/services/open-api/model/article-res-dto.ts
/**
 * Trend.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ArticleResDto { 
    id?: string | null;
    created?: string;
    title?: string | null;
    text?: string | null;
    url?: string | null;
    pageSource?: string | null;
    typeName?: string | null;
    typeId?: number;
    searchWordId?: string | null;
    searchWord?: string | null;
    searchWordImage?: string | null;
}